organized event files into own controllers and views folders/files
