+++ app/controllers/Application.java
@@ -191,78 +191,4 @@ public class Application extends Controller {
     renderText("");
   }
 
-  public static void events(Long id) {
-		User user = id == null ? user() : (User) User.findById(id);
-		render(user);
-  }
-	
-  public static void addEvent() {
-		render();
-  }
-
-public static void addEventInvite(Long eventId, Long guestId){
-		Event event = Event.findById(eventId);
-		User guest = User.findById(guestId);
-		event.newEventInvite(guest);
-	}
-	
-	public static void deleteEvent(Long eventId, Long page) {
-		Event event = Event.findById(eventId);
-		event.delete();
-		news(page);
-	}
-	
-	
-	public static void editEvent() {
-		//Event event = Event.findById(id);
-		render();
-	}
-
-	public static void event_create(Event curEvent) {
-		User currentUser = user();
-		Event newEvent = new Event();
-
-		validation.required(curEvent.eventName).message("Event name is required");
-		validation.required(curEvent.eventScript).message("Event description is required");
-		validation.required(curEvent.eventLocation).message("Event location is required");
-		//validation.required(curEvent.startDate).message("Event start date and time are required");		
-		//validation.required(curEvent.open || curEvent.friends || curEvent.invite).message("Event privacy is required");
-
-		if (validation.hasErrors()) {
-			Event thisEvent = curEvent;
-			renderTemplate("Application/addEvent.html", thisEvent);
-		} else {
-			Event event = newEvent;
-			event.author = currentUser;
-
-			event.eventName = curEvent.eventName;
-			event.eventScript = curEvent.eventScript;
-			event.eventLocation = curEvent.eventLocation;
-			System.out.println(curEvent.eventName);
-			/*
-			event.startDate = curEvent.startDate;
-			if (given(curEvent.endDate)) {
-				event.endDate = curEvent.endDate;
-			}
-			*/
-//			if (given(curEvent.open)) {
-//				event.open = TRUE;
-//			}
-//			if (given(curEvent.friends)) {
-//				event.friends = TRUE;
-//			}
-//			if (given(curEvent.invite)) {
-//				event.invite = TRUE;
-//			}
-
-			event.save();
-			//this should pull up to an extended events page
-			//add a Locatiom
-			//invite guests
-			//upload event photos
-			//upload newsfeed
-			
-			editEvent();
-		}
-	}
 }
