+++ app/views/Application/news.html
@@ -8,7 +8,7 @@
     <br>
     <a href="@{Application.news()}">news</a>
     <br>
-    <a href="@{Application.events()}">Events</a>
+    <a href="@{Events.events()}">Events</a>
     <br>
     <a href="@{Application.about()}">about</a>
     <br>
