+++ app/views/Events/editEvent.html
@@ -1,0 +1,2 @@
+#{extends 'main.html' /}
+To Do: modify event:
