+++ app/views/Events/events.html
@@ -1,0 +1,4 @@
+#{extends 'main.html' /}
+To Do: should list events:
+<a class=button href="@{Events.addEvent()}">+ Create Event</a>
+
